syn-test: IUB-side -- fixed power-up issue
This was due to a missing synchronizer flip-flop on the
SHIFT signal. The SHIFT signal may drift from time to time, which leads to an
uncertainty in detecting a falling edge on this signal. Since we are using the
falling edge to shift in bits from the IUB, a wrong detection of the falling
edge can lead to a wrong value of the shifted bit by effectively missing bits.

Adding the synchronizer flip-flop fixes the issue.