Changed code for tests in Japan
Our Japanese colleagues do not have several of the boards in the PoGO+ system,
therefore the code has been modified to not take into account the IUB-side
signals. All power enable signals are hard-coded and the rest of the bits sent
by the IUB, so all the IUB signals will be optimized away on synthesis.

The current version only enables one FADC and the SpaceWire Router and the
SpaceCube, as in the current test setup at KTH the DIO is connected on the Power
Board to the SpaceCube power enable line.

Even with no Power Board in Japan, the fact the DIO and SpaceWire Router are
enabled does not affect operation.